requirements.txt
Django>=1.8,<1.9
SQLAlchemy==0.8.1
argparse==1.2.1
django-debug-toolbar==0.9.4
django-debug-toolbar-template-timings
django-extensions==1.1.1
django-markdown-deux==1.0.5
django-postgrespool==0.2.4
django-redis==3.3
djangorestframework==3.1.0
factory-boy==2.1.1
gunicorn==17.5
httpretty==0.6.1
ipython==0.13.2
markdown2==2.1.0
mock==1.0.1
psycopg2==2.5
pytz==2013b
redis==2.7.6
requests==2.6.0
requests-oauthlib==0.8.0
six==1.3.0
twitter==1.9.4
